pyproject.toml
[tool.poetry]
name = "quark"
version = "0.1.0"
description = ""
authors = ["Jerry <zhangji87@gmail.com>"]
readme = "README.md"
packages = [{include = "quark"}]

[tool.poetry.extras]
gunicorn = ["gunicorn"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.5"
pymysql = "^1.0.3"
flask-login = "^0.6.2"
marshmallow = "^3.19.0"
gunicorn = {version = "^20.1.0", optional = true}
python-dateutil = "^2.8.2"
sqlalchemy = "^2.0.20"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.270"
mypy = "^1.3.0"
types-python-dateutil = "^2.8.19.13"
pytest = "^7.3.2"
freezegun = "^1.2.2"

[tool.ruff]
select = ["E", "F", "W", "Q", "COM", "I", "N", "B", "RUF"]
line-length = 100
target-version = "py310"

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.ruff.flake8-quotes]
inline-quotes = "single"

[[tool.mypy.overrides]]
module = [
    "flask_login",
]
ignore_missing_imports = true
